requirements.txt
tqdm==4.66.4
pandas==2.0.3
scikit-learn==1.3.2
anthropic==0.31.2
openai==1.37.1
reka-api>=2.0.0
google-generativeai==0.3.1
mistralai==0.4.2
datasets==2.20.0
tabulate==0.9.0
